+++ example/android/app/src/main/kotlin/com/example/icon_generator_test/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.icon_generator_test
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}