+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1
+
+* Fixed an issue causing icons to have incorrect width.
+* Added example project.
+
 # 0.1.0
 
 * Initial development release.